phase3/requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pytest>=8.0.0
